Add Typescript's types to Flow Instances page
Move Flow Instances page to typescript
